20220413/func/requirements.txt
flask
google-cloud-firestore
google-cloud-pubsub
